# docs/openapi/openapi-sdce-1.yaml
openapi: 3.0.1
info:
  contact:
    email: onap-discuss@lists.onap.org
    name: ONAP
    url: https://onap.readthedocs.io
  description: SDC API for onboarding (SDCE-1)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: "SPC API: SDCE-1"
  version: "1.0"
servers:
- description: SDCE-1 APIs
  url: /sdc
paths:
...
  /v1.0/externaltesting/endpoints/{endpointId}/executions/{executionId}:
    post:
      operationId: getExecution
      parameters:
      - in: path
        name: endpointId
        required: true
        schema:
          type: string
      - in: path
        name: executionId
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
  /v1.0/externaltesting/endpoints/{endpointId}/scenarios:
    get:
      operationId: getScenarios
      parameters:
      - in: path
        name: endpointId
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
  /v1.0/externaltesting/endpoints/{endpointId}/scenarios/{scenario}/testcases:
    get:
      operationId: getTestcases
      parameters:
      - in: path
        name: endpointId
        required: true
        schema:
          type: string
      - in: path
        name: scenario
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
  /v1.0/externaltesting/endpoints/{endpointId}/scenarios/{scenario}/testsuites:
    get:
      operationId: getTestsuites
      parameters:
      - in: path
        name: endpointId
        required: true
        schema:
          type: string
      - in: path
        name: scenario
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
  /v1.0/externaltesting/endpoints/{endpointId}/scenarios/{scenario}/testsuites/{testsuite}/testcases/{testcase}:
    get:
      operationId: getTestcase
      parameters:
      - in: path
        name: endpointId
        required: true
        schema:
          type: string
      - in: path
        name: scenario
        required: true
        schema:
          type: string
      - in: path
        name: testsuite
        required: true
        schema:
          type: string
      - in: path
        name: testcase
        required: true
        schema:
          type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
  /v1.0/externaltesting/executions:
    get:
      operationId: getValidationResult
      parameters:
      - in: query
        name: requestId
        schema:
          type: string
      - in: query
        name: endPoint
        schema:
          type: array
          items:
            type: string
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
    post:
      operationId: execute
      parameters:
      - in: query
        name: vspId
        schema:
          type: string
      - in: query
        name: vspVersionId
        schema:
          type: string
      - in: query
        name: requestId
        schema:
          type: string
      - in: header
        name: USER_ID
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
      responses:
        default:
          content:
            application/json: {}
          description: default response
      tags:
      - SDCE-1 APIs
      - External-Testing
...
components:
  schemas:
...
    Attachment:
      type: object
      properties:
        contentDisposition:
          $ref: '#/components/schemas/ContentDisposition'
        contentId:
          type: string
        contentType:
          type: object
          properties:
            parameters:
              type: object
              additionalProperties:
                type: string
            subtype:
              type: string
            type:
              type: string
            wildcardSubtype:
              type: boolean
            wildcardType:
              type: boolean
        dataHandler:
          type: object
          properties:
            allCommands:
              type: array
              items:
                type: object
                properties:
                  commandClass:
                    type: string
                  commandName:
                    type: string
            commandMap:
              type: object
              properties:
                mimeTypes:
                  type: array
                  items:
                    type: string
              writeOnly: true
            content:
              type: object
            contentType:
              type: string
            dataSource:
              type: object
              properties:
                contentType:
                  type: string
                inputStream:
                  type: object
                name:
                  type: string
                outputStream:
                  type: object
            inputStream:
              type: object
            name:
              type: string
            outputStream:
              type: object
            preferredCommands:
              type: array
              items:
                type: object
                properties:
                  commandClass:
                    type: string
                  commandName:
                    type: string
            transferDataFlavors:
              type: array
              items:
                type: object
                properties:
                  defaultRepresentationClassAsString:
                    type: string
                  flavorJavaFileListType:
                    type: boolean
                  flavorRemoteObjectType:
                    type: boolean
                  flavorSerializedObjectType:
                    type: boolean
                  flavorTextType:
                    type: boolean
                  humanPresentableName:
                    type: string
                  mimeType:
                    type: string
                  mimeTypeSerializedObject:
                    type: boolean
                  primaryType:
                    type: string
                  representationClassByteBuffer:
                    type: boolean
                  representationClassCharBuffer:
                    type: boolean
                  representationClassInputStream:
                    type: boolean
                  representationClassReader:
                    type: boolean
                  representationClassRemote:
                    type: boolean
                  representationClassSerializable:
                    type: boolean
                  subType:
                    type: string
        headers:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
        object:
          type: object
...
    ContentDisposition:
      type: object
      properties:
        filename:
          type: string
        parameters:
          type: object
          additionalProperties:
            type: string
        type:
          type: string
...